requirements/base.txt
# Django
django==1.9.13

# Configuration
django-environ==0.4.4


# Time zones support
pytz==2017.2


google-api-python-client==1.6.4
django-braces==1.11.0

coverage==4.4.1
django-coverage-plugin==1.5.0

Sphinx==1.6.5
django-extensions==1.9.7
Werkzeug==0.12.2
django-test-plus==1.0.20
factory-boy==2.9.2

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.9.0